Extract CityGML building attribute parser
